# apis/dolibarr/openapi.yaml
openapi: 3.0.3
info:
  title: Dolibarr REST API
  description: |
    API REST de Dolibarr ERP/CRM. Les appels se font sur l'URL de base de votre instance
    (ex. https://votre-domaine.fr/api/index.php).
    Activer le module « API REST » dans Configuration → Modules/Applications.
    La clé API (DOLAPIKEY) se génère dans le profil utilisateur (onglet Clé API).
  version: 1.0.0
  contact:
    name: Dolibarr
    url: https://www.dolibarr.org
  license:
    name: GPL
    url: https://www.gnu.org/licenses/gpl.html
servers:
  - url: https://bdcloud.luxgreen.net/api/index.php
    description: Instance exemple (remplacer par votre URL)
  - url: /api/index.php
    description: URL relative (base à configurer)
...
security:
  - ApiKeyHeader: []
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: DOLAPIKEY
      description: Clé API générée dans le profil utilisateur (onglet Clé API)
    EntityHeader:
      type: apiKey
      in: header
      name: DOLAPIENTITY
      description: "(Optionnel) Entité en multi-société. '' = première société."
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "404"
            message:
              type: string
...
    InvoiceLine:
      type: object
      properties:
        fk_product:
          type: string
        label:
          type: string
        desc:
          type: string
        qty:
          type: number
        subprice:
          type: number
        tva_tx:
          type: number
    InvoiceCreate:
      type: object
      required:
        - socid
        - lines
      properties:
        socid:
          type: string
          description: ID du tiers (client)
        date:
          type: integer
          description: Timestamp
        lines:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLine'
...
paths:
...
  /invoices:
    get:
      tags: [Factures]
      summary: Liste des factures
      operationId: listInvoices
      parameters:
        - name: sortfield
          in: query
          schema:
            type: string
        - name: sortorder
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
        - name: limit
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Liste des factures
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '403':
          description: Accès refusé
        '500':
          description: Erreur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Factures]
      summary: Créer une facture
      operationId: createInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCreate'
      responses:
        '200':
          description: ID de la facture créée
          content:
            application/json:
              schema:
                type: integer
        '400':
          description: Données invalides (socid, lines requis)
        '403':
          description: Accès refusé
        '500':
          description: Erreur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /invoices/{id}:
    get:
      tags: [Factures]
      summary: Détail d'une facture
      operationId: getInvoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Détail de la facture
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Facture non trouvée
...